Bugfix: handle null correctly

+++ src/main/java/com/aliyun/odps/jdbc/utils/transformer/ByteTransformer.java
@@ -7,7 +7,7 @@ public class ByteTransformer extends AbstractTransformer {
   @Override
   public Object transform(Object o, String charset) throws SQLException {
     if (o == null) {
-      return 0;
+      return (byte) 0;
     }
 
     if (Number.class.isInstance(o)) {
