+++ src/main/java/com/aliyun/odps/jdbc/utils/transformer/FloatTransformer.java
@@ -7,7 +7,7 @@ public class FloatTransformer extends AbstractTransformer {
   @Override
   public Object transform(Object o, String charset) throws SQLException {
     if (o == null) {
-      return 0;
+      return (float) 0;
     }
 
     if (Number.class.isInstance(o)) {
